split tests into unit and integration, with separate Make targets
