+++ Makefile
@@ -13,7 +13,8 @@ lint:
 start:
 	pipenv run python run.py
 
-TEST_TARGET=tests
+UNIT_TESTS=tests/unit
+INTEGRATION_TESTS=tests/integration
 
 docker-test: REDIS_PORT=6379
 docker-test: test
@@ -23,5 +24,8 @@ check:
 # TODO: re-enable as soon as possible.
 #	pipenv check
 
-test: check lint
-	APP_SETTINGS=TestingConfig pipenv run pytest $(TEST_TARGET) --cov frontstage --cov-report term-missing	
+unit-tests: check lint
+	APP_SETTINGS=TestingConfig pipenv run pytest $(UNIT_TESTS) --cov frontstage --cov-report term-missing	
+
+integration-tests: check lint
+	APP_SETTINGS=TestingConfig pipenv run pytest $(INTEGRATION_TESTS) --cov frontstage --cov-report term-missing
