+++ tests/integration/test_availability_message.py
