+++ tests/integration/test_error_handlers.py
@@ -6,7 +6,7 @@ from requests.exceptions import ConnectionError
 
 from frontstage import app
 from frontstage.exceptions.exceptions import ApiError, JWTValidationError
-from tests.app.mocked_services import url_get_respondent_email, url_oauth_token, party, encoded_jwt_token
+from tests.integration.mocked_services import url_get_respondent_email, url_oauth_token, party, encoded_jwt_token
 
 
 class TestErrorHandlers(unittest.TestCase):
