+++ tests/integration/test_passwords.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 from frontstage.common import verification
 from config import TestingConfig
 from frontstage import app
-from tests.app.mocked_services import token, url_password_change, url_reset_password_request, \
+from tests.integration.mocked_services import token, url_password_change, url_reset_password_request, \
     url_verify_token, url_get_respondent_by_email
 
 encoded_valid_email = 'ImV4YW1wbGVAZXhhbXBsZS5jb20i.vMOqeMafWQpuxbUBRyRs29T0vDI'
