+++ tests/integration/test_session.py
