+++ tests/integration/views/surveys/test_access_survey.py
@@ -3,8 +3,8 @@ import unittest
 from unittest.mock import patch
 
 from frontstage import app
-from tests.app.mocked_services import (collection_exercise, collection_instrument_seft, survey, business_party,
-                                       encoded_jwt_token, encrypted_enrolment_code)
+from tests.integration.mocked_services import (collection_exercise, collection_instrument_seft, survey, business_party,
+                                               encoded_jwt_token, encrypted_enrolment_code)
 
 
 class TestAccessSurvey(unittest.TestCase):
