+++ tests/integration/views/surveys/test_surveys_list.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import patch
 
 from frontstage import app
-from tests.app.mocked_services import encoded_jwt_token, survey_list_history, survey_list_todo
+from tests.integration.mocked_services import encoded_jwt_token, survey_list_history, survey_list_todo
 
 
 class TestSurveyList(unittest.TestCase):
