+++ tests/integration/views/surveys/test_upload_survey.py
@@ -8,7 +8,7 @@ from structlog import wrap_logger
 
 from frontstage import app
 from frontstage.exceptions.exceptions import CiUploadError
-from tests.app.mocked_services import business_party, case, encoded_jwt_token, survey, url_upload_ci
+from tests.integration.mocked_services import business_party, case, encoded_jwt_token, survey, url_upload_ci
 
 logger = wrap_logger(logging.getLogger(__name__))
